requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-google-genai==0.0.5
openai==1.6.1
google-generativeai==0.3.2
python-pptx==0.6.23
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
python-dotenv==1.0.0
